pyproject.toml
[tool.poetry]
name = "ochen-polezniy-sait"
version = "0.1.0"
description = "Очень полезный сайт"
authors = ["OlegShilov-neLox <olegshilov00@gmail.com>"]
license = "free"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
pillow = "^11.0.0"
django-simple-captcha = "^0.6.0"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
